+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Drop support for ArangoDB <= 3.5 (versions already EOL)
 - Add new admin commands to fetch agency dump and agency state
 - Add Graceful shutdown as finalizer (supports kubectl delete)
+- Add Watch to Lifecycle command
 
 ## [1.2.4](https://github.com/arangodb/kube-arangodb/tree/1.2.4) (2021-10-22)
 - Replace `beta.kubernetes.io/arch` Pod label with `kubernetes.io/arch` using Silent Rotation